Discord/requirements.txt
discord.py

fastapi

pydantic-settings

uvicorn

hypercorn